feat: implement responsive header component with logo and authentication
- Add Header component with DragMe logo and authentication buttons
- Implement responsive design with mobile-first approach
- Add dark/light mode support for logo and text colors
- Position logo on left and auth buttons on right with proper spacing
- Prevent viewport overflow with flexible layout and button protection
- Include CSS modules for scoped styling and maintainability
- Support logo text hiding on mobile devices for better UX
- Integrate with existing authentication context and components
