+++ src/index.css
@@ -1,68 +1,47 @@
+@import url('https://fonts.googleapis.com/css2?family=Roboto:ital,wght@0,100..900;1,100..900&display=swap');
 :root {
-  font-family: system-ui, Avenir, Helvetica, Arial, sans-serif;
-  line-height: 1.5;
-  font-weight: 400;
-
-  color-scheme: light dark;
-  color: rgba(255, 255, 255, 0.87);
-  background-color: #242424;
-
-  font-synthesis: none;
-  text-rendering: optimizeLegibility;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
+  --text: #f9f9f9;
+  --background: #202020;
+  --buttonBackground: #1a1a1a;
+  --buttonBorder: #d1d5db;;
+  --inputBackground: #1a1a1a;
+  --inputBorder: #202020;
+  --googleButtonBackground: #1a1a1a;
+  --cardBackground: #1a1a1a;
+  --cardSpanBackground: #1a1a1a;
+  --spanTextColor: #f9f9f9;
+  --dividerColor: #2a2b2d;
+  font-family: 'Roboto', sans-serif;
 }
-
-a {
-  font-weight: 500;
-  color: #646cff;
-  text-decoration: inherit;
-}
-a:hover {
-  color: #535bf2;
-}
-
-body {
-  margin: 0;
-  display: flex;
-  place-items: center;
-  min-width: 320px;
-  min-height: 100vh;
-}
-
-h1 {
-  font-size: 3.2em;
-  line-height: 1.1;
+body{
+  color: var(--text);
+  background-color: var(--background);
 }
 
 button {
+  color: var(--text);
   border-radius: 8px;
   border: 1px solid transparent;
   padding: 0.6em 1.2em;
   font-size: 1em;
   font-weight: 500;
   font-family: inherit;
-  background-color: #1a1a1a;
+  background-color: var(--buttonBackground);
+  border-color: var(--borderColor);
   cursor: pointer;
-  transition: border-color 0.25s;
-}
-button:hover {
-  border-color: #646cff;
-}
-button:focus,
-button:focus-visible {
-  outline: 4px auto -webkit-focus-ring-color;
 }
 
 @media (prefers-color-scheme: light) {
   :root {
-    color: #213547;
-    background-color: #ffffff;
-  }
-  a:hover {
-    color: #747bff;
-  }
-  button {
-    background-color: #f9f9f9;
+    --text: #1a1a1a;
+    --buttonBorder: #202020;
+    --background: #ffffff;
+    --buttonBackground: #f9f9f9;
+    --inputBackground: #f9f9f9;
+    --cardSpanBackground: white;
+    --cardBackground: white;
+    --googleButtonBackground: #f9f9f9;
+    --spanTextColor: #6b7280;
+    --dividerColor: #e5e7eb;
   }
 }
